pyproject.toml
[project]
name = "bazar-analysis"
version = "0.1.0"
description = "Bazaar endgame board extraction pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "curl-cffi>=0.15.0",
    "duckdb>=1.2.2",
    "httpx>=0.28.1",
    "imagehash>=4.3.1",
    "numpy>=2.2.5",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "rapidfuzz>=3.13.0",
    "typer>=0.15.3",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.52.0",
]

[project.scripts]
bazar-analysis = "bazar_analysis.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bazar_analysis"]
